requirements.txt
matplotlib==3.6.2
numpy==1.23.5
Pillow==9.3.0
tqdm==4.64.1
wandb==0.13.10
rasterio==1.3.4
shapely==2.0.1
hydra-core==1.3.2
omegaconf==2.3.0
torchinfo==0.0.1
torchmetrics==1.0.1